requirements.txt
# FLIR Grasshopper 相機預覽程式所需的Python套件

# OpenCV - 用於圖像處理和顯示
# 注意：使用4.8.x版本以避免與numpy的兼容性問題
opencv-python==4.8.1.78

# NumPy - 用於數組處理（OpenCV的依賴）
# 注意：必須使用1.x版本，2.x版本與opencv-python 4.8.x不兼容
numpy<2

# PyYAML - 用於讀取配置文件
pyyaml>=6.0

# 注意：PySpin 需要從FLIR官網單獨下載和安裝
# 下載地址：https://www.flir.com/products/spinnaker-sdk/
# 選擇對應的作業系統版本並安裝PySpin Python綁定
